requirements.txt
aiohttp>=3.9.0
pytest-aiohttp>=1.0.5
pytest-asyncio>=0.23.0
